:dizzy: /api/group → routes/api/group.js
