+++ lib/utils.js
@@ -53,6 +53,11 @@ module.exports = {
 		return req.headers['x-forwarded-for'] || req.connection.remoteAddress;
 	},
 
+	slug_match: (slug) => {
+		const slug_regex = /^[a-z][a-z0-9_-]+$/i;
+		return slug.match(slug_regex) ? true : false;
+	},
+
 	error_page: (rsp, type) => {
 		rsp.render('page', {
 			title: 'Error',
